refactor: 컴포넌트 디렉토리 분리
